Allow frontend to be served as static content
This doesn't change existing behavior, but allows the frontend to
be served as static content by setting the FRONTEND_BUILD_DIR
environment variable to the React build path.

This is likely more of a hack than a proper solution, but implements
the final missing piece for the local deployment scripts to work.